VMM/CPUMR3-x86: Fixed a release log statement printing the incorrect size of mapped MTRR region due to missing typecast.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@110305 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/VMMR3/target-x86/CPUMR3-x86.cpp
@@ -3252,7 +3252,7 @@ static int cpumR3MapMtrrs(PVM pVM)
             pCtxMsrs->msr.MtrrFix4K_F0000  = 0x0505050505050505;
             pCtxMsrs->msr.MtrrFix4K_F8000  = 0x0505050505050505;
         }
-        LogRel(("CPUM: Mapped %Rhcb (%u bytes) of RAM using fixed-range MTRRs\n", _1M, _1M));
+        LogRel(("CPUM: Mapped %Rhcb (%u bytes) of RAM using fixed-range MTRRs\n", (uint64_t)_1M, _1M));
     }
     else
     {
